features: сделал навигацию в профиле.
