+++ src/components/pages/profile/profile.module.css
@@ -1,14 +1,9 @@
-.form {
-  display: flex;
-  width: 480px;
-  flex-direction: column;
-  align-items: center;
-}
-.button {
-  margin-bottom: 80px;
-}
 .link {
   text-decoration: none;
+}
+
+.activeLink {
+  color: white;
 }
 .linkContainer {
   display: flex;
@@ -17,16 +12,7 @@
   height: 192px;
   margin-bottom: 80px;
 }
-.button2 {
-  text-decoration: none;
-  color: #4c4cff;
-  border: none;
-  background: none;
-  cursor: pointer;
-}
-.linkFormContainer {
-  align-self: flex-end;
-}
+
 .leftContentContainer {
   width: 320px;
 }
