+++ src/components/profile-form/profile-form.module.css
@@ -1,0 +1,46 @@
+.form {
+    display: flex;
+    width: 480px;
+    flex-direction: column;
+    align-items: center;
+  }
+  .button {
+    margin-bottom: 80px;
+  }
+  /* .link {
+    text-decoration: none;
+  }
+  
+  .activeLink {
+    color: white;
+  }
+  .linkContainer {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    height: 192px;
+    margin-bottom: 80px;
+  } */
+  .button2 {
+    text-decoration: none;
+    color: #4c4cff;
+    border: none;
+    background: none;
+    cursor: pointer;
+  }
+  .linkFormContainer {
+    align-self: flex-end;
+  }
+  .leftContentContainer {
+    width: 320px;
+  }
+  /* .profileContainer {
+    display: flex;
+    justify-content: space-between;
+    max-width: 860px;
+    margin: 180px auto 0 auto;
+  }
+  .text {
+    color: #8585ad;
+    opacity: 0.4;
+  } */
